Refactor GeminiApi constructor to accept modelName parameter
The GeminiApi constructor in api.ts has been modified to accept an additional parameter called modelName. This allows for more flexibility in specifying the generative model to be used. The modelName parameter is now passed to the getGenerativeModel method of the GoogleGenerativeAI class. This change improves the configurability of the GeminiApi class.